make Register No decimal

+++ PetaDB/dbo/Tables/RptRegister.sql
@@ -18,5 +18,5 @@
     [GST] DECIMAL(12, 2) NULL, 
 	[EncashCertDetails] VARCHAR(100) NULL,
 	[Remarks] VARCHAR(100) NULL,
-	[RegisterNo] INT NULL, 
+	[RegisterNo] DECIMAL(10, 2) NULL, 
     )
